# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CRM module for Uralintern
  description: API documentation for Uralintern CRM Module
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: kayakto server url
security:
  - BearerAuth: []
...
paths:
...
  /events/update/{eventId}:
    put:
      tags:
        - event-controller
      operationId: updateEvent
      description: "Обновление данных мероприятия"
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/hide/{eventId}:
    put:
      tags:
        - event-controller
      operationId: hideEvent
      description: "Поменять статус мероприятия на \"скрытое\""
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events/start-event:
    post:
      tags:
        - event-controller
      operationId: startEvent
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int64
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/post:
    post:
      tags:
        - event-controller
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /auth/register:
    post:
      tags:
        - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /auth/register-with-token:
    post:
      tags:
        - auth-controller
      operationId: registerWithToken
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /auth/refresh:
    post:
      tags:
        - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /auth/login:
    post:
      tags:
        - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /messages/event/{eventId}:
    get:
      tags:
        - message-controller
      operationId: getEventMessages
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/{eventId}:
    get:
      tags:
        - event-controller
      operationId: getEventByID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/my-events/{userId}:
    get:
      tags:
        - event-controller
      operationId: getMyEvents
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/condition/{eventId}:
    get:
      tags:
        - event-controller
      operationId: getEventConditionByID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/admin/{adminId}:
    get:
      tags:
        - event-controller
      operationId: getEventsByAdminId
      parameters:
        - name: adminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events/active:
    get:
      tags:
        - event-controller
      operationId: getOpenedEvents
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events/delete/{eventId}:
    delete:
      tags:
        - event-controller
      operationId: deleteEventByID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  responses:
      OkResponse:
        description: OK, запрос выполнен успешно
        content:
          '*/*':
            schema:
              type: boolean
      BadRequestResponse:
        description: Bad Request, запрос неверный со стороны клиента
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      NotFoundResponse:
        description: Not Found, такого элемента не существует
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      InternalServerErrorResponse:
        description: Internal Server Error, запрос вызвал ошибку на сервере
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "User with email email123@gmail.com not found"
...
    EventRequest:
      type: object
      properties:
        title:
          type: string
        descriptionText:
          type: string
        adminId:
          type: integer
          format: int64
        managerId:
          type: integer
          format: int64
        eventStartDate:
          type: string
          format: date-time
          example: "2024-11-01T22:00:00+05:00"
        eventEndDate:
          type: string
          format: date-time
          example: "2024-11-01T22:00:00+05:00"
        enrollmentStartDate:
          type: string
          format: date-time
          example: "2024-11-01T22:00:00+05:00"
        enrollmentEndDate:
          type: string
          format: date-time
          example: "2024-11-01T22:00:00+05:00"
        chatUrl:
          type: string
        numberSeatsStudent:
          type: integer
          format: int32
        condition:
          type: string
          enum:
            - PREPARATION
            - REGISTRATION_OPEN
            - NO_SEATS
            - REGISTRATION_CLOSED
            - IN_PROGRESS
            - FINISHED
            - HIDDEN
            - DELETED
          example: "REGISTRATION_OPEN"
...
    UserRequest:
      required:
        - email
        - firstName
        - lastName
        - sign
        - telegramUrl
        - vkUrl
      type: object
      properties:
        firstName:
          type: string
          example: "Николай"
        lastName:
          type: string
          example: "Иванов"
        surname:
          type: string
          example: "Евгеньевич"
        telegramUrl:
          type: string
          example: "https://t.me/kayakto"
        vkUrl:
          type: string
          example: "https://vk.com/kayakto"
        role:
          type: string
          enum:
            - ADMIN
            - CURATOR
            - MANAGER
            - STUDENT
          example: "STUDENT"
        competencies:
          type: string
          example: "Spring Boot, Gradle, git, Java, Docker"
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: "mail123@mail.ru"
        password:
          type: string
          example: "password"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT